requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pymongo[srv]==4.4.1
dnspython==2.4.2
python-dotenv==1.0.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.24.3
scikit-learn==1.3.0
huggingface-hub==0.20.3
transformers==4.36.2
bcrypt==4.0.1
flask-jwt-extended==4.5.3
email-validator==2.1.0
flask-mail==0.9.1
celery==5.3.4
redis==5.0.1
flask-limiter==3.5.0
flask-caching==2.1.0

# Existing packages remain...

# NEW packages for Task 07:
flask-login==0.7.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.24.3 